# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Stock Management",
        "description": "API for managing stock items, categories, and sales",
        "contact": {
            "email": "inolatse@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/enter/article": {
            "post": {
                "tags": [
                    "EnterStore"
                ],
                "summary": "Add a single item to the stock",
                "operationId": "88586379c3174cef7a4fb4f00ac5ed89",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "quantity",
                                    "article_id"
                                ],
                                "properties": {
                                    "quantity": {
                                        "description": "Quantity of the article being entered",
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "article_id": {
                                        "description": "ID of the article",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item successfully added to the stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EnterStore"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/enter/articles": {
            "post": {
                "tags": [
                    "EnterStore"
                ],
                "summary": "Add multiple items to the stock",
                "operationId": "052889a52c3592f3a4b6b398938c6267",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "description": "List of articles to be entered",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "quantity",
                                                "article_id"
                                            ],
                                            "properties": {
                                                "quantity": {
                                                    "description": "Quantity of the article",
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "article_id": {
                                                    "description": "ID of the article",
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Items successfully added to the stock",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Entrées de stock enregistrés avec succès."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/EnterStore"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/v1/enter/articles/statistics": {
            "get": {
                "tags": [
                    "EnterStore"
                ],
                "summary": "Get overall statistics for entered stock",
                "operationId": "91c2124c4a89c017ad60b65ddcfeb9c6",
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_quantity": {
                                            "description": "Total quantity of all entered items",
                                            "type": "integer",
                                            "example": 100
                                        },
                                        "articles": {
                                            "description": "List of articles with their total entered quantity",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "article_id": {
                                                        "description": "ID of the article",
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "article_name": {
                                                        "description": "Name of the article",
                                                        "type": "string",
                                                        "example": "Article A"
                                                    },
                                                    "quantity": {
                                                        "description": "Total quantity entered for the article",
                                                        "type": "integer",
                                                        "example": 50
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "most_enter": {
                                            "description": "The most entered article",
                                            "properties": {
                                                "article_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "article_name": {
                                                    "type": "string",
                                                    "example": "Article A"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/api/v1/enter/articles/statistics/period": {
            "get": {
                "tags": [
                    "EnterStore"
                ],
                "summary": "Get stock entry statistics filtered by date range",
                "operationId": "3434600369e39d4dd31648122c0f2dd1",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for filtering (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for filtering (format: YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-11-30"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_quantity": {
                                            "description": "Total quantity of all entered items in the selected date range",
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "articles": {
                                            "description": "List of articles with their total entered quantity in the date range",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "article_id": {
                                                        "description": "ID of the article",
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "article_name": {
                                                        "description": "Name of the article",
                                                        "type": "string",
                                                        "example": "Article B"
                                                    },
                                                    "quantity": {
                                                        "description": "Total quantity entered for the article",
                                                        "type": "integer",
                                                        "example": 30
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "most_enter": {
                                            "description": "The most entered article in the date range",
                                            "properties": {
                                                "article_id": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "article_name": {
                                                    "type": "string",
                                                    "example": "Article B"
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 30
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EnterStore": {
                "required": [
                    "quantity",
                    "article_id"
                ],
                "properties": {
                    "quantity": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "article_id": {
                        "type": "integer",
                        "format": "int32"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}